pyproject.toml
[tool.poetry]
name = "omni-arb"
version = "0.1.0"
description = "پروژه ترکیبی برای یکپارچه‌سازی FinGPT، FinRL و کانکتور MEXC فیوچرز"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
FinGPT = { path = "external/FinGPT", develop = true }
FinRL = { path = "external/FinRL", develop = true }
fastapi = "^0.95.0"
uvicorn = "^0.18.0"
pydantic = "^1.10.0"
httpx = "^0.24.0"
aiohttp = "^3.8.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
